Adds two static variable to PrefixQuery to globally control expansions
PrefixQuery can be quite aggressive when expanding short prefixes. This might result in an unusable query
which has more than 1024 OR clauses). Therefore these variables can limit the total expansions and
also impose a relation between the prefix length and the length of the expanded token. I.e.
it might be wise to only expand to a token which is three times longer, but not to one which is 15 times longer.
This is especially useful for artificially created tokens which combine many words.

By default both limitations are turned off so that the change is backwards compatible.